# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Astra VI 3D Cadastre API",
    "description": "SIH26011 spatial workspace and governance prototype. Proposed 3D ULPIN is a prototype extension, not a government standard.",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/evidence": {
      "post": {
        "tags": [
          "evidence"
        ],
        "summary": "Add Evidence",
        "operationId": "add_evidence_api_v1_evidence_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvidenceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "evidence"
        ],
        "summary": "Evidence",
        "operationId": "evidence_api_v1_evidence_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "object_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Object Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvidenceOut"
                  },
                  "title": "Response Evidence Api V1 Evidence Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rights": {
      "post": {
        "tags": [
          "rights"
        ],
        "summary": "Add Right",
        "operationId": "add_right_api_v1_rights_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RightCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RightOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "rights"
        ],
        "summary": "Rights",
        "operationId": "rights_api_v1_rights_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "object_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Object Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RightOut"
                  },
                  "title": "Response Rights Api V1 Rights Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ulpin/{object_id}": {
      "post": {
        "tags": [
          "ulpin"
        ],
        "summary": "Identity",
        "operationId": "identity_api_v1_ulpin__object_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Object Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ulpin": {
      "get": {
        "tags": [
          "ulpin"
        ],
        "summary": "Find Identity",
        "operationId": "find_identity_api_v1_ulpin_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "identifier",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100,
              "title": "Identifier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IdentityOut"
                  },
                  "title": "Response Find Identity Api V1 Ulpin Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews": {
      "post": {
        "tags": [
          "reviews"
        ],
        "summary": "Review",
        "operationId": "review_api_v1_reviews_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "Reviews",
        "operationId": "reviews_api_v1_reviews_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewOut"
                  },
                  "title": "Response Reviews Api V1 Reviews Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{review_id}": {
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "Review Detail",
        "operationId": "review_detail_api_v1_reviews__review_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reviews/{review_id}/decisions": {
      "post": {
        "tags": [
          "reviews"
        ],
        "summary": "Decide",
        "operationId": "decide_api_v1_reviews__review_id__decisions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/audit": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "History",
        "operationId": "history_api_v1_audit_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "entity_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditOut"
                  },
                  "title": "Response History Api V1 Audit Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "actor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Actor Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "string",
            "format": "uuid",
            "title": "Entity Id"
          },
          "detail": {
            "additionalProperties": true,
            "type": "object",
            "title": "Detail"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "actor_id",
          "action",
          "entity_type",
          "entity_id",
          "detail",
          "created_at"
        ],
        "title": "AuditOut"
      },
...
      "DecisionCreate": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "ACCEPTED_PROTOTYPE",
              "RETURNED"
            ],
            "title": "Decision"
          },
          "reason": {
            "type": "string",
            "maxLength": 3000,
            "minLength": 5,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision",
          "reason"
        ],
        "title": "DecisionCreate"
      },
      "DecisionOut": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "ACCEPTED_PROTOTYPE",
              "RETURNED"
            ],
            "title": "Decision"
          },
          "reason": {
            "type": "string",
            "maxLength": 3000,
            "minLength": 5,
            "title": "Reason"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "review_id": {
            "type": "string",
            "format": "uuid",
            "title": "Review Id"
          },
          "reviewer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Reviewer Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision",
          "reason",
          "id",
          "review_id",
          "reviewer_id",
          "created_at"
        ],
        "title": "DecisionOut"
      },
      "EvidenceCreate": {
        "properties": {
          "object_id": {
            "type": "string",
            "format": "uuid",
            "title": "Object Id"
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          },
          "description": {
            "type": "string",
            "maxLength": 3000,
            "minLength": 3,
            "title": "Description"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "object_id",
          "dataset_id",
          "description"
        ],
        "title": "EvidenceCreate"
      },
      "EvidenceOut": {
        "properties": {
          "object_id": {
            "type": "string",
            "format": "uuid",
            "title": "Object Id"
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          },
          "description": {
            "type": "string",
            "maxLength": 3000,
            "minLength": 3,
            "title": "Description"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "object_id",
          "dataset_id",
          "description",
          "id",
          "created_at",
          "created_by"
        ],
        "title": "EvidenceOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IdentityOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "object_id": {
            "type": "string",
            "format": "uuid",
            "title": "Object Id"
          },
          "issued_geometry_id": {
            "type": "string",
            "format": "uuid",
            "title": "Issued Geometry Id"
          },
          "identifier": {
            "type": "string",
            "title": "Identifier"
          },
          "label": {
            "type": "string",
            "const": "Proposed 3D ULPIN",
            "title": "Label"
          },
          "scheme_version": {
            "type": "string",
            "title": "Scheme Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "object_id",
          "issued_geometry_id",
          "identifier",
          "label",
          "scheme_version",
          "created_at"
        ],
        "title": "IdentityOut"
      },
...
      "ReviewCreate": {
        "properties": {
          "geometry_id": {
            "type": "string",
            "format": "uuid",
            "title": "Geometry Id"
          },
          "validation_job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Validation Job Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "geometry_id",
          "validation_job_id"
        ],
        "title": "ReviewCreate"
      },
      "ReviewOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "object_id": {
            "type": "string",
            "format": "uuid",
            "title": "Object Id"
          },
          "geometry_id": {
            "type": "string",
            "format": "uuid",
            "title": "Geometry Id"
          },
          "validation_job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Validation Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "submitted_by": {
            "type": "string",
            "format": "uuid",
            "title": "Submitted By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "object_id",
          "geometry_id",
          "validation_job_id",
          "status",
          "submitted_by",
          "created_at"
        ],
        "title": "ReviewOut"
      },
      "RightCreate": {
        "properties": {
          "object_id": {
            "type": "string",
            "format": "uuid",
            "title": "Object Id"
          },
          "evidence_id": {
            "type": "string",
            "format": "uuid",
            "title": "Evidence Id"
          },
          "right_type": {
            "type": "string",
            "enum": [
              "RECORDED_OWNERSHIP",
              "SHARED_USE",
              "ACCESS",
              "EASEMENT",
              "RESTRICTION"
            ],
            "title": "Right Type"
          },
          "party_reference": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Party Reference"
          },
          "description": {
            "type": "string",
            "maxLength": 3000,
            "minLength": 3,
            "title": "Description"
          },
          "is_synthetic": {
            "type": "boolean",
            "title": "Is Synthetic",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "object_id",
          "evidence_id",
          "right_type",
          "party_reference",
          "description"
        ],
        "title": "RightCreate"
      },
      "RightOut": {
        "properties": {
          "object_id": {
            "type": "string",
            "format": "uuid",
            "title": "Object Id"
          },
          "evidence_id": {
            "type": "string",
            "format": "uuid",
            "title": "Evidence Id"
          },
          "right_type": {
            "type": "string",
            "enum": [
              "RECORDED_OWNERSHIP",
              "SHARED_USE",
              "ACCESS",
              "EASEMENT",
              "RESTRICTION"
            ],
            "title": "Right Type"
          },
          "party_reference": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Party Reference"
          },
          "description": {
            "type": "string",
            "maxLength": 3000,
            "minLength": 3,
            "title": "Description"
          },
          "is_synthetic": {
            "type": "boolean",
            "title": "Is Synthetic",
            "default": false
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "object_id",
          "evidence_id",
          "right_type",
          "party_reference",
          "description",
          "id",
          "created_at",
          "created_by"
        ],
        "title": "RightOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}